fix: add value props to Select components and extract useSemesterAutofill hook
